requirements.txt
click==8.1.8
cryptography==44.0.2
numpy==2.2.4
Pillow==11.1.0
pycryptodome==3.22.0
setuptools==77.0.3
